sessionCounter and expiredSessions declares as long instead of int.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@934337 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/session/StandardSession.java
@@ -793,11 +793,11 @@ public class StandardSession
                 if (timeAlive > manager.getSessionMaxAliveTime()) {
                     manager.setSessionMaxAliveTime(timeAlive);
                 }
-                int numExpired = manager.getExpiredSessions();
+                long numExpired = manager.getExpiredSessions();
                 numExpired++;
                 manager.setExpiredSessions(numExpired);
                 int average = manager.getSessionAverageAliveTime();
-                average = ((average * (numExpired-1)) + timeAlive)/numExpired;
+                average = (int) (((average * (numExpired-1)) + timeAlive)/numExpired);
                 manager.setSessionAverageAliveTime(average);
             }
